1、优化iOS／OSX的重载方法筛选规则 2、调整Unity下 iOS ／OSX下最新的导出类型方式（注：Android尚未调整，请不要进行master分支的clone）
Former-commit-id: 33cc0625fe119da0af63df727a74ebbfa367c0d7

+++ Source/iOS_OSX/LuaScriptCoreTests-iOS/Person.m
@@ -29,9 +29,9 @@ static LSCManagedValue *_managedFunc = nil;
     NSLog(@"%@ %ld years old", self.name, (long)age);
 }
 
-- (void)speakWithPos:(double)pos
+- (void)speakWithPos:(int)pos
 {
-    NSLog(@"%@ pos = %f", self.name, pos);
+    NSLog(@"%@ pos = %d", self.name, pos);
 }
 
 - (void)dealloc
